File generation checks prerequisites
